Feature/admin page - add password protection to delete feature

+++ pages/api/trays/index.js
@@ -18,13 +18,21 @@ export default async function handler(req, res) {
       break;
 
     case 'DELETE':
-      const deleted = await db.collection('trays').deleteMany({});
-      res.json({
-        success: deleted.acknowledged,
-        message: deleted.deletedCount
-          ? `All ${deleted.deletedCount} trays have been deleted`
-          : 'No trays to delete.'
-      });
+      try {
+        if (req.body.delete === 'G00dBye~') {
+          const deleted = await db.collection('trays').deleteMany({});
+          res.json({
+            success: deleted.acknowledged,
+            message: deleted.deletedCount
+              ? `All ${deleted.deletedCount} trays have been deleted`
+              : 'No trays to delete.'
+          });
+        } else req.status(500);
+      } catch (err) {
+        res
+          .status(500)
+          .json({ success: false, message: 'Incorrect password.' });
+      }
       break;
 
     default:
